chore: refactor code 4
