+++ packages/amplify-cli/src/__tests__/commands/gen2-migration/generate/amplify/data/data.renderer.test.ts
@@ -1,6 +1,18 @@
 import { GraphqlApi } from '@aws-sdk/client-appsync';
-import { DataRenderer, DataRenderOptions } from '../../../../../../commands/gen2-migration/generate/amplify/data/data.renderer';
-import { ClassifiedVtlFiles, classifyVtlFiles } from '../../../../../../commands/gen2-migration/generate/amplify/data/data.generator';
+import {
+  DataRenderer,
+  DataRenderOptions,
+  groupExtendedResolverFiles,
+  computeSpliceIndexes,
+  PIPELINE_3_SLOT_MAP,
+  PIPELINE_4_SLOT_MAP,
+  ExtendedResolverFile,
+} from '../../../../../../commands/gen2-migration/generate/amplify/data/data.renderer';
+import {
+  ClassifiedVtlFiles,
+  classifyVtlFiles,
+  ParsedExtended,
+} from '../../../../../../commands/gen2-migration/generate/amplify/data/data.generator';
 import { TS } from '../../../../../../commands/gen2-migration/generate/ts';
 
 jest.unmock('fs-extra');
@@ -20,6 +32,135 @@ function renderResolverOutput(classifiedResolvers: ClassifiedVtlFiles): string {
   const nodes = renderer.render(opts);
   return TS.printNodes(nodes);
 }
+
+describe('groupExtendedResolverFiles', () => {
+  const makeExtended = (typeName: string, fieldName: string, slot: string, order: number, templateType: 'req' | 'res'): ParsedExtended => ({
+    kind: 'extended',
+    typeName,
+    fieldName,
+    slot,
+    order,
+    templateType,
+    filename: `${typeName}.${fieldName}.${slot}.${order}.${templateType}.vtl`,
+  });
+
+  it('groups entries by typeName.fieldName', () => {
+    const entries: ParsedExtended[] = [
+      makeExtended('Mutation', 'createTodo', 'init', 1, 'req'),
+      makeExtended('Query', 'listItems', 'postAuth', 1, 'req'),
+    ];
+    const result = groupExtendedResolverFiles(entries);
+    expect(result.size).toBe(2);
+    expect(result.has('Mutation.createTodo')).toBe(true);
+    expect(result.has('Query.listItems')).toBe(true);
+  });
+
+  it('sorts within group by slot pipeline execution order', () => {
+    const entries: ParsedExtended[] = [
+      makeExtended('Mutation', 'createTodo', 'finish', 1, 'req'),
+      makeExtended('Mutation', 'createTodo', 'init', 1, 'req'),
+      makeExtended('Mutation', 'createTodo', 'auth', 1, 'req'),
+    ];
+    const result = groupExtendedResolverFiles(entries);
+    const groups = result.get('Mutation.createTodo')!;
+    expect(groups.map((g) => g.slot)).toEqual(['init', 'auth', 'finish']);
+  });
+
+  it('sorts within same slot by numeric order', () => {
+    const entries: ParsedExtended[] = [
+      makeExtended('Mutation', 'createTodo', 'init', 3, 'req'),
+      makeExtended('Mutation', 'createTodo', 'init', 1, 'req'),
+      makeExtended('Mutation', 'createTodo', 'init', 2, 'req'),
+    ];
+    const result = groupExtendedResolverFiles(entries);
+    const groups = result.get('Mutation.createTodo')!;
+    expect(groups.map((g) => g.order)).toEqual([1, 2, 3]);
+  });
+
+  it('pairs req and res templates for same slot+order', () => {
+    const entries: ParsedExtended[] = [
+      makeExtended('Mutation', 'createTodo', 'init', 1, 'req'),
+      makeExtended('Mutation', 'createTodo', 'init', 1, 'res'),
+    ];
+    const result = groupExtendedResolverFiles(entries);
+    const groups = result.get('Mutation.createTodo')!;
+    expect(groups).toHaveLength(1);
+    expect(groups[0].reqFile).toBe('Mutation.createTodo.init.1.req.vtl');
+    expect(groups[0].resFile).toBe('Mutation.createTodo.init.1.res.vtl');
+  });
+
+  it('handles entry with only req file (no res)', () => {
+    const entries: ParsedExtended[] = [makeExtended('Mutation', 'createTodo', 'init', 1, 'req')];
+    const result = groupExtendedResolverFiles(entries);
+    const groups = result.get('Mutation.createTodo')!;
+    expect(groups).toHaveLength(1);
+    expect(groups[0].reqFile).toBe('Mutation.createTodo.init.1.req.vtl');
+    expect(groups[0].resFile).toBeUndefined();
+  });
+
+  it('handles entry with only res file (no req)', () => {
+    const entries: ParsedExtended[] = [makeExtended('Mutation', 'createTodo', 'init', 1, 'res')];
+    const result = groupExtendedResolverFiles(entries);
+    const groups = result.get('Mutation.createTodo')!;
+    expect(groups).toHaveLength(1);
+    expect(groups[0].reqFile).toBeUndefined();
+    expect(groups[0].resFile).toBe('Mutation.createTodo.init.1.res.vtl');
+  });
+});
+
+describe('computeSpliceIndexes', () => {
+  const makeResolverFile = (slot: string, order: number): ExtendedResolverFile => ({
+    typeName: 'Mutation',
+    fieldName: 'createTodo',
+    slot,
+    order,
+  });
+
+  it('uses 3-function pipeline for Query', () => {
+    const resolverFiles: ExtendedResolverFile[] = [{ typeName: 'Query', fieldName: 'getTodo', slot: 'init', order: 1 }];
+    const result = computeSpliceIndexes('Query', 'getTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_3_SLOT_MAP['init']);
+  });
+
+  it('uses 3-function pipeline for Subscription', () => {
+    const resolverFiles: ExtendedResolverFile[] = [{ typeName: 'Subscription', fieldName: 'onCreateTodo', slot: 'init', order: 1 }];
+    const result = computeSpliceIndexes('Subscription', 'onCreateTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_3_SLOT_MAP['init']);
+  });
+
+  it('uses 3-function pipeline for delete Mutation (fieldName starts with "delete")', () => {
+    const resolverFiles: ExtendedResolverFile[] = [{ typeName: 'Mutation', fieldName: 'deleteTodo', slot: 'init', order: 1 }];
+    const result = computeSpliceIndexes('Mutation', 'deleteTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_3_SLOT_MAP['init']);
+  });
+
+  it('uses 4-function pipeline for create Mutation', () => {
+    const resolverFiles: ExtendedResolverFile[] = [{ typeName: 'Mutation', fieldName: 'createTodo', slot: 'init', order: 1 }];
+    const result = computeSpliceIndexes('Mutation', 'createTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_4_SLOT_MAP['init']);
+  });
+
+  it('uses 4-function pipeline for update Mutation', () => {
+    const resolverFiles: ExtendedResolverFile[] = [{ typeName: 'Mutation', fieldName: 'updateTodo', slot: 'init', order: 1 }];
+    const result = computeSpliceIndexes('Mutation', 'updateTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_4_SLOT_MAP['init']);
+  });
+
+  it('computes correct splice index with running offset for multiple entries', () => {
+    const resolverFiles: ExtendedResolverFile[] = [makeResolverFile('init', 1), makeResolverFile('auth', 1), makeResolverFile('finish', 1)];
+    const result = computeSpliceIndexes('Mutation', 'createTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(PIPELINE_4_SLOT_MAP['init'] + 0);
+    expect(result.entries[1].spliceIndex).toBe(PIPELINE_4_SLOT_MAP['auth'] + 1);
+    expect(result.entries[2].spliceIndex).toBe(PIPELINE_4_SLOT_MAP['finish'] + 2);
+  });
+
+  it('computes init.2 and finish.1 on create mutation correctly', () => {
+    const resolverFiles: ExtendedResolverFile[] = [makeResolverFile('init', 2), makeResolverFile('finish', 1)];
+    const result = computeSpliceIndexes('Mutation', 'createTodo', resolverFiles);
+    expect(result.entries[0].spliceIndex).toBe(1);
+    expect(result.entries[1].spliceIndex).toBe(5);
+  });
+});
 
 describe('DataRenderer - resolver code generation', () => {
   describe('override resolvers only', () => {
